@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
-    "@angular/cdk": "^13.1.1",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
-    "@angular/material": "^13.1.1",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
-    "@angular/service-worker": "~13.2.0",
+    "@angular/animations": "^15.2.4",
+    "@angular/cdk": "^15.2.4",
+    "@angular/common": "^15.2.4",
+    "@angular/compiler": "^15.2.4",
+    "@angular/core": "^15.2.4",
+    "@angular/forms": "^15.2.4",
+    "@angular/material": "^15.2.4",
+    "@angular/platform-browser": "^15.2.4",
+    "@angular/platform-browser-dynamic": "^15.2.4",
+    "@angular/router": "^15.2.4",
+    "@angular/service-worker": "^15.2.4",
     "@commitlint/cli": "^17.0.2",
     "@commitlint/config-conventional": "^17.0.2",
     "@ngx-translate/core": "^14.0.0",
@@ -30,7 +30,7 @@
     "bootstrap": "^5.1.3",
     "commitizen": "^4.2.4",
     "husky": "^8.0.1",
-    "ngx-spinner": "^13.1.1",
+    "ngx-spinner": "^15.0.1",
     "prettier": "^2.7.0",
     "rxjs": "~7.5.0",
     "swiper": "^8.3.1",
@@ -38,9 +38,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.6",
-    "@angular/cli": "~13.2.6",
-    "@angular/compiler-cli": "~13.2.0",
+    "@angular-devkit/build-angular": "^15.2.4",
+    "@angular/cli": "^15.2.4",
+    "@angular/compiler-cli": "^15.2.4",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
@@ -49,7 +49,7 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.8.4"
   },
   "config": {
     "commitizen": {
